requirements.txt
amqp==5.0.6
asgiref==3.3.4
autopep8==1.5.7
beautifulsoup4==4.9.3
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
crispy-bootstrap5==0.4
Django==3.2.4
django-bootstrap5==2.0.0
django-celery-results==2.0.1
django-crispy-forms==1.11.2
Faker==8.6.0
importlib-metadata==2.1.1
kombu==5.1.0
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pycodestyle==2.7.0
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.10.0.0
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
